# perf(turbopack): Update rustc-hash from 1.1.0 to 2.1.0