--- a/Cargo.toml
+++ b/Cargo.toml
@@ -182,7 +182,7 @@ rayon = "1.10.0"
 regex = "1.10.6"
 reqwest = { version = "=0.11.17", default-features = false }
 rstest = "0.16.0"
-rustc-hash = "1.1.0"
+rustc-hash = "2.1.0"
 semver = "1.0.16"
 serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0.138"
@@ -201,6 +201,7 @@ smallvec = { version = "1.13.1", features = [
 ] }
 sourcemap = "9.0.0"
 strsim = "0.11.1"
+swc-rustc-hash = { package = "rustc-hash", version = "1.1.0" } # used with swc
 syn = "1.0.107"
 tempfile = "3.3.0"
 thread_local = "1.1.8"