--- a/turbopack/crates/turbopack-ecmascript/tests/tree-shaker/analyzer/2/output.md
+++ b/turbopack/crates/turbopack-ecmascript/tests/tree-shaker/analyzer/2/output.md
@@ -403,16 +403,16 @@ graph TD
     Export(
         "external1",
     ): 0,
-    Exports: 14,
+    Export(
+        "external2",
+    ): 11,
     Export(
         "foo",
     ): 13,
     Export(
         "foobar",
     ): 11,
-    Export(
-        "external2",
-    ): 11,
+    Exports: 14,
 }
 ```
 
@@ -631,16 +631,16 @@ import "__TURBOPACK_PART__" assert {
     Export(
         "external1",
     ): 0,
-    Exports: 12,
+    Export(
+        "external2",
+    ): 10,
     Export(
         "foo",
     ): 4,
     Export(
         "foobar",
     ): 11,
-    Export(
-        "external2",
-    ): 10,
+    Exports: 12,
 }
 ```