--- a/turbopack/crates/turbopack-ecmascript/tests/tree-shaker/analyzer/typeof-1/output.md
+++ b/turbopack/crates/turbopack-ecmascript/tests/tree-shaker/analyzer/typeof-1/output.md
@@ -184,10 +184,10 @@ graph TD
 ```
 {
     ModuleEvaluation: 7,
-    Exports: 8,
     Export(
         "GET",
     ): 6,
+    Exports: 8,
 }
 ```
 
@@ -301,10 +301,10 @@ import "__TURBOPACK_PART__" assert {
 ```
 {
     ModuleEvaluation: 7,
-    Exports: 8,
     Export(
         "GET",
     ): 6,
+    Exports: 8,
 }
 ```